added check for missing or extra mods to launch
and changed a whole bunch of other things that didn't really change the
operation of the program at all...Added a weird generator-based
setup-thread-handler to qt_launch. Added a couple methods that are
wrappers around the db connection as a context manager to DatabaseMan.
Delayed loading of profile data until requested by the UI instead
of in the init of ModMan. Moved bunch of stuff out of the init of
ModManWindow (thus the setup-thread-thing, mentioned above).