Added custom API for updating profiles combobox
i.e., I added one function to use instead of the three or four that
would have been required to get the box to update correctly when
relying on QT's regular signal/slot mechanisms. Now the model and view
are both up to date after closing the 'New Profile' popup dialog. The
problem lay in the fact that modifiying the items in a combobox
programatically (rather than using the line-editor) doesn't seem to be
expected behavior...Or maybe I just don't get it. But this works and
is all that is needed.