added name-edit, check if table needs saving
Basically if the user modifies a row--either by changing a mod's
display name or toggling the enabled-checkbox--the model now checks to
see the row that was just edited has been edited before, and, if so,
does it now match its initial state before the first edit. If the mod
has been set back to that initial state (i.e. the user manually undid
their changes to the name or enabled status) and no other mod-entries
have been changed, the save/cancel buttons will be disabled again. This
is mainly to avoid unnecessary disk-writes and to not annoy the user
with 'unsaved changes' notifications if all they had done was
accidentally toggle a mod, then corrected their error.

Unfortunately, now using the "Revert" button breaks this whole
system...but that's next.