DB fld-order less hardcoded; install order not PK
By making the ordinality of the mods a regular column (and not an alias
for the rowid), it should now be easier/possible to reorder the
mods-list (aka the install-order) as desired during program execution.
When saved to disk, the mods will be sorted on their ordinal and saved
to the file as such, but without their ordinal rank number. When loaded
back in, they will receive an auto-incrementing ordinal rank based on
the order they are read from the save-file (or pulled from the
mods-directory, if the profile is missing at program start).