Replaced QFSModel subclass with custom QAIModel SC
The QFileSystemModel was definitely functional and...very featureful.
Full of features. Most of which I realized I really didn't need. And it
required programming _around_ the data I needed rather than integrating
it directly. Basically there was way more overhead than necessary for
this task and I couldn't figure out how to get the control over the
data that _was_ necessary for the task.  So I scrapped the whole thing
and implemented a custom model, subclassing from QAbstractItemModel and
building on top of the regular python filesystem-interaction tools.
This is more than sufficient as I'm not really going to be manipulating
the FS in any way, I just needed to project a view of what's on it.
Handling the tree view, checkmark states, etc., was a bit more
challenging, but following some of the Qt tree-model examples and
looking into how they actually implemented parts of QTreeWidget in c++
helped to resolve those issues rather nicely.