Added tree to hold list of hidden files
Mainly intended for sending to the db or saving to json.  It takes
advantage of the defaultdict() to create an autovivifying tree.
It requires using a special insert function that takes a list
of key names, but the function also takes care of placing any 'leaves'
aka non-directory files into a special list under their parent
directory node.  This allows same-level access to the both the files
and folders in a directory and makes the dict/json representation look
very clean and meaningful.