impl. func to load paths of all modfiles from HD
It was fast on my SSD with a small simulation mod setup; I'll
definitely need to test how it fairs on a spinning HDD with a real
heavy mod load before committing to using this method. Also need to
analyze the performance of the database with thousands of entries. (is
that possible? Or are all BIG mods packed up in bsa's and esp's now?)