Hidden files are now properly tracked by DB
Also, the 'algorithm' I was using to iterate through the file hierarchy
has been replaced by something far more elegant, readable, and
recursive. It's a slow process, but my ultimate dream is for this code
to look like it was written by someone who is not, in fact, criminally
insane.