Don't know why they broke, but ftree checks fixed
The cascading checkmarks in the mod file tree suddenly stopped...well,
cascading. Un/checking  A directory that contained other directories would
not cause the checkmarks of any of its children to to change
appropriately; I was sure this used to be working. It may be due to the
filter...Well anyway, I managed to get it working again, but it
definitely feels a fair bit more sluggish than before... I'll call that
a todo.