Tried to find cause of slowdown on some mod moves.
specifically, when mods are shifted up or down towards the bottom of
the list, there's a slight (but noticeable) lag before the update
occurs that doesn't happen towards the top of the list. I didn't really
have any luck finding the cause, though.  Thought maybe it was due to
copying and replacing all those namedtuples, so I converted ModEntry to
a mutable class with __slots__ (with some methods that mimic namedtuple's),
but that didn't seem to make much (if any) difference. And now I've got
that, and I'm not sure it was a good idea. Well, anyway, the delay is
small, so that will have to go on the backburner for now.  If it's due
to something in Qt...I doubt I'll ever figure it out.