Simplified data() and index() methods
There was a relatively expensive check in index() that was never satisfied,
so I took it out to simplify and speed up the call. And I replaced my
function/lambda calls in data() with just a few cascading if/elifs,
significantly reducing the number of method calls made per second by
the program; especially seeing as how data() was called nearly 1 million
times during ~30 seconds using the program...yeah, that method can take
all the optimization it can get.