Improved performance of row movement
Was calling some expensive methods multiple times to perform checks
that could have been made much more cheaply. Which is how they are now.

Still doesn't **entirely** fix the lag when moving stuff around towards the
bottom of the list, but it is noticeably better, which means the delay
itself has gone from 'notieable' to 'just perceptible'. Hopefully
that'll be good enough and not too many people will notice.