fancified list of mod names on filetree tab
'Active Mods Only' button works now, and label changes accordingly, as
well as showing how many mods are visible out of the total amount
installed.

And speaking of fancy, I discovered a feature of the mod table I didn't
even know existed: having the table in focus and typing a letter will
scroll to the next row found containing that letter (it'll be at the
bottom of the table's frame).  This is kinda like a builtin name
search--and I got it for free! I'll still definitely want to
implement a better search feature, but this is neat for now.