added implementation of timed undo-cmd compression
It's a subclass of the undo.py Stack class that runs a Timer after
every append, restarting the timer if another append happens before the
timer runs out, and grouping together any commands that occur while the
timer is running.  The grouped commands will be done/undone as a single
command, just like the regular group() context manager. Infact, groups
made with group() can be added to this timed group as a single unit,
effectively nesting the group levels.

I'm impressed and pleased that the implementation works pretty much
exactly as it should. However, something doesn't feel right about it...
the code seems somewhat awkward and clunky, and the whole thing feels
fragile.  Maybe something will come to me, or maybe it's good enough.
I'll just have to see if it has a tendency to break randomly.