I decided disabling unchecked files was silly
Well, only files that were underneath an unchecked directory. But it
was initially mainly for the visual clue that the whole dir was hidden,
but I think the checkboxes are good enough for that. And having
everything disabled was kind of annoying if you then wanted to recheck
a single item under a disabled directory...