consolidated signal connections; fixed tree filter
Most of the signal, action, button, slot, etc. connections have been
consolidated into setup methods dedicated to just connecting things,
rather than being scattered around in the various other setup
functions. I just found this easier to grok.

Also, I managed to bypass the default filtering behaviour on the tree
model (that is, where an item's parents must all match the filter for
the item to be shown) so that only the file names are considered for
filtering: If a folder contains a file (at any depth below it) that
matches the filter, that folder will show up in the filtered tree.
Folders without matching files will of course be hidden.