Trying to use db transactions for saving ftree
Well, trying to get the save/revert buttons to apply to the file tree
tab if the user in the file tree tab, and the mod table tab if they're
there instead. For the file viewer, I'm trying to mark (but not commit)
changes to the hidden files in the database; hitting save will commit
to the DB and write to disk, while hitting reset will rollback to the
last saved state. Additionally, the save buttons should react and be
enabled/disabled accordingly.

Right now...it's kinda working. Well, it's not crashing horribly, so I
consider that a good first step.