and now it works right.
I made a QtDesigner plugin for the fun of it: a simple extension of
QlineEdit that just emits a signal when you hit escape. Kinda neat that
it works. Anyway, the signal is used to collapse and clear the search
box when the user hits escape with it focused. Also, I've changed it so
that the search text is saved (and only updated) when the user hits
enter, so it isn't directly tied to the search box text anymore. This
way, even if the box has been cleared or collapsed, one can still hit
find-next or find-previous to repeat the most recent search.