modmanager is now a module, not a class
I realized that having an instance of the main ModManager was fairly
superfluous, since the main point of the class was simply to proxy
commands on to the various other managers. It occurred to me that
having module level functions (and global references to the actual
instanced managers) would work just as well with probably less
overhead, and would remove the need to pass the manager instance around
through constructors. Any module that needs access to the manager can
now simply 'import modmanager' and call the methods directly.