All attempts at using libarchive were disastrous.
Some of it was due to bugs in the library; much of it was likely due to
me. But either way, trying to get a solution to handling multiple types
of archives with a single interface kept leading to more and more
workarounds and convoluted code. So I threw it all out and rewrote an
un-archiver module based entirely around 7zip. And it works great. No
wonder it's shipped with ModOrganizer. Even though it requires
subprocesses and asyncio, etc., nothing else came anywhere close to
being as easy or--more importantly--correct to use. So I guess that'll
be a new requirement, but hey: it's not like it's unprecedented.