extract after fomod-installer working now...
...well, almost. The files are extracted to correct directory (the real
mod directory), and the progress bar seems to work accurately, but the
files are not placed in the correct destinations: the source folders are
currently being copied in whole. Unlike unrar, it doesn't seem that 7z
has an option for removing paths from the output files, so the files
will need to be rearranged manually (i.e outside of the 7z command)
after extraction.