created a case-insensitive faux-filesystem module
Because I hate being happy. And, uh, also because it will hopefully
help with representing the structure of mod archives without having to
actually unpack them entirely or try to 'fake' folder-structure with
nested dictionaries or some such.

The FS is pretty simple and bare bones, but is based around a
bastardized version of PurePosixPath that I rigged up to make case-
insensitive (which seems to work perfectly, at least in simple tests
so far). It uses tables of inodes<->paths and only supports one link
per inode, but since the "files" aren't intended to actually hold data
(their names are the only data I'm concerned about), I don't think
hard/sym-links are likely to be a desired feature.

This will likely require building a new TreeModel to display the
filesystem structure in the manual-install dialog...see first sentence
of these comments for details on that.