Tweak undo buttons; actions now work almost right.
View is not updated correctly after a rename operation--specifically,
when the name of a parent directory changes, that dir is reordered to
the correct sorted location, but its contents do not always follow it.
This does seem to be entirely a visual bug; the underlying fs tree is
in fact updated as it should be.