Trying to implement a managed dir-merge
By managed, I mean that if a conflict occurs during the operation, the
user can decided what to do and the merger will continue where it left
off. User should even be able to cancel the operation in the middle of
the process and whatever progress was achieved up to that point will be
encapsulated into a single, complete undo action.  I think a lot of the
pieces are in place, but...I haven't tested it, and I'm very tired and
it was beginning to make my brain feel twisty, so I wouldn't expect it
to work all that well just yet...will definitely need to review and
test thoroughly.