ManagerWindow: Improve code structure
Somewhat. Trying to make it less of a 'follow the tangled thread' game
when looking for where a function is defined/used.

Also added pyqtSlot() decorators to most of the functions used as
slots. According to the docs, connecting to a decorated slot uses less
memory and is slightly faster, which never hurts. And it makes the
purpose of those methods more explicit.