Add more comments
I'm relieved that I left as many comments as I did when first writing
this; however, there's still plenty of areas that are uncommented and
of course entirely baffling to me now that it's months later. So as I
figure things out, I'm trying to fill in the documentation gaps.