Waffle
I was going to make an AppSettings class (like there used to be) and
subclass QObject so that it could send a signal when a preference was
changed, but I decided not to. After I'd done the conversion. And I
didn't really feel like changing it all back. So now it's a
singleton object for not much good reason.