Replace timedundo with QUndoStack
And it works, too. Mostly. QUndo* still seems to group clumps of
similar commands together whether or not you told it to merge or
macro them. Probably not a big issue. Appears to be snappy and
accurate, as well (as accurate as it was, anyway). There's a bit of
problem with some things not knowing the correct clean/unsaved state:
particularly the revert/save buttons are enabled precisely when they
should be disabled, and vice versa. Probably just need to switch a
'not' somewhere.

Finally, looks like it ought to be fairly simple to plug the
file-tree-viewer's undo stuff in there too without any extra hassle.
There's significantly less manual book-keeping to do for this setup,
which is a relief.