Move some DB queries to DBManager
I'm trying to keep all the SQL confined to that one class/module, so
that if there's a problem with database access, I don't have to hunt
down the offending SQL code (which just looks like any other string)
all over the codebase. The issue, of course, is that the class is
getting a bit bloated with all the specific method calls...but the
alternative just feels leaky. There's probably an _alternative_
alternative that's better than both that I haven't thought of yet.