Prevent Manager creation on import
Several modules attempted to access the ModManager instance at the
module level when they were imported; if that Manager instance didn't
exist, it's programmed to be created, ensuring that access won't fail.
This wasn't normally an issue, but basically caused the entire backend
to start up when the module was imported elsewhere OUTSIDE the
application environment, like, say, in QtDesigner.