FileTree: Implement working* savepoint-based undo
*As usual, though, "working" comes with a caveat. Undo/redo seems to
work as intended while clicking around in the tree of a single
mod--including setting and updating the checkstates of the affected
items automatically! But, when switching mods by clicking on a
different item in the list on the left, the changes are committed to
the database and the undostack cleared. This may not be an issue as
no data is lost, but a user may expect to be able to make changes, look
at a different mod, and then come back and still be able to undo the
changes the made to the first one (y'know, automagically with ctrl-z,
not manually). I'll have to decide on a policy for that.