Track correctable errors with Alert objects
The alert object is just a namedtuple that contains information about
the problem that occurred and what steps the user can take to correct
the issue. It also includes a method to check if the issue is still
active. This data-object should make it easier to pass information
about missing files or other fixable errors around the application and
to display that information to the user.