Implement draft of view-alerts button
Managed to get a couple things to work. The key was using a
QWidgetAction with a plain QToolButton on the toolbar, rather than
trying to have a QAction popup a window. With this, I initially had it
just drop down a text box where I set the text from the list of alerts.
For something a little more fancy (and more sustainable, I think) I
changed the widget to a QListView using a QStringListModel where I set
the model's list from a composition of the active Alert objects. These
should be fairly straightforward to scale up to something more involved
like a TreeView to show just the label, but when expanded will show the
problem description and suggested fix.