Create Master Control Program, Submanager class
In an attempt to keep things orderly, all sub-managers will now be
beholden to the master manager. They will not be created until the
master manager deems it necessary, and they will only interact with
the other managers through the main manager, also known as the
Master Control Program (or MCP). Surely, this cannot go wrong.