Handle unset directories w/ exception
Rather than keeping track of valid directories and checking, we now
raise and catch an exception if the given directory is unset. Should
keep things a bit cleaner and more performant. The exception can be
ignored if desired to return a false-like value.

Also added more granularity to the process of setting/changing the
active profile.