Fix previous "cleverness"
Specifically, I tried to avoid dropping and regenerating the table of
mods when the mods-folder did not change on a profile-switch, instead
just updating the table with info from the new profile. That was a very
bad idea, leading to violated db-constraints, invalid mod listings, and
lots of headaches. Now returned to the drop-and-regen policy.

However, I believe that I have managed (at least partially) to prevent
spuriously re-examining the disk for the list of ALL files from ALL
installed mods when switching profiles. It should only happen when the
mod-storage directory is changed. Might still be some kinks to work out.