Add 'Base' version of dbmanager
As in, a base class that contains only the generic database-interaction
stuff. Having both that and the app-specific code in one class was
getting unwieldy. May also add base versions of some/all of the other
managers.

This change also involved a bit of refactoring of all the managers to
better handle cooperative-subclassing.