Clean out utils/__init__.py
It was a hodge-podged mess, and a lot of the things were used in just
one place in the project. Those pieces have been consolidate to where
they are needed until there is a need to use them elsewhere, in which
case they will likely be added to submodule.

__init__.py no longer imports anything from its submodules either. They
must be imported directly. This is to prevent importing a lot of
unnecessary things when only a very specific part of the package is
desired.