Add initial version of QObject wrapper for modmgr
At the moment, this just wraps the methods that manipulate the set of
active alerts to check for actual changes and emit a signal if any
occurred. I think there is definitely room to expand this class and
emit signals/create slots for many other mod-manager actions,
simplifying the interaction between the GUI and the backend.