Naively replace ModManager w/ QModManager in GUI
It's probably broken as heck, and it doesn't yet take advantage of any
of the signal/slot mechanics. Looks like it will take some doing to
truly decouple the interface from the data. I should probably segregate
a lot of the methods in ManagerWindow into 'data-dependent' or
'UI-only' groups (or something along those lines).