Mess with alerts-view to make all text visible
QLabels were cutting off the bottom of some of the descriptions; for
some reason, setting the text to be selectable by mouse enlarged the
label enough to see everything--everything and then some. There's some
blank space now, but we may have to live with it.

Also force the menu and the treewidget to resize on every alert-change
to have enough (and only enough) room to see all the expanded alerts.
This method will be problematic if there are a large number of issues
to show, though...may need to think of a different way.