Fix (kinda) adding files from base Data dir
I.e. 'Skyrim/Data'. And yes, while the 'fake' mods for these do show
up, they're by default added to the end of the mod list (opposite where
they should be). Their component files don't appear on the mod-files
tab because they're not in the 'proper' mod directory. Attempting to
change the name/order of these 'mods' will surely cause all sorts of
issues since then they'd be in the profile's modinfo.json AS WELL as
read in from the files on disk...

Do we need a SEPARATE db table/config file to handle these? Should we
add an 'unmanaged' boolean field to indicate that these are not found
in the regular mods directory? Should I just give up and set my
computer on fire? I'll have to think about this...