Make files directly accessible from mod object
Can access them as either a list or a tree (a utils.tree.AutoTree,
that is). This was used to replace the old system of building the
FSItems for the fileviewer by actually reading the directories on disk;
now, the entries are pulled from the ModEntry object, which itself gets
its data directly from the 'modfiles' database table. This table is
populated every time the program is loaded and when a mod is
installed/removed.