Interface: Add module for handling profiles
Instead of doing it all in the main window. The class is created, but
not yet integrated. This also exposed some problems in current profile
handling (e.g. renaming a profile does not update the selector model,
making the 'renamed' profile appear to not exist until the app is
restarted, while the old one may be recreated if the user fiddles with
the profile selector some more).