Make it work?
I must be dreaming, because it seems that it's actually working
again...main window comes up, table is populated with all the correct
fields, files are visible in the File tab...just like it's supposed to
be!

Admittedly, I've yet to test most (er, any) of the advanced
functionality like changing mod order, hiding files, or saving data. I
know that not all of that has been entirely updated yet. But the fact
that it gets this far is a pleasant suprise!