Change 'move' algorithm, add prepare_move()
After realizing that I'd been treating the collection like a list
rather than a conglomeration of dicts, I saw a far simpler way to
implement the move operation. As a bonus, it makes undo-ing a move much
simpler, as well.

Also added a prepare_move() method that returns a callable object which
'stores' a move operation to execute at a later time. It also has an
'undo' mode. Clearly intended to feed the QUndoManager for
move-operations in the mod table.