Reimplement row-movement actions
Can rearrange the mod order again, even using drag-and-drop. The whole
process is far simpler now, I believe. Also, the command building and
pushing is handled on the view as it should be, rather than on the model,
taking a lot complexity out of the model workings. That should be done
for the other commands that are currently pushed by the model, as well.

It still crashes when using the 'move-to-bottom' action (i.e. trying to
move an item or items to the end of the mod collection). It's an
indexerror that shouldn't be hard to fix once I track it down.