Fix double-undo/redo bug
Seems that the generated undo/redo actions of the UndoManager
automatically call undo()/redo() on the active stack...no need to call
it manually. Which makes sense when you think about it...which
obviously I didn't do. Anyway, yeah, because of that, every time the
action was triggered two undos/redo would happen. But not anymore.