Improving hidden-files handling
Trying to get rid of (or at least lessen) the amount of work that goes
into every single hide/unhide-file action. Doing that by indexing the
list of files for the mod into a flat list and performing a binary
search (rather than linear) when looking for files to hide/unhide.

Also an attempt to move the undo actions out of the model.