Fix Fomod installation...mostly
The installer no longer crashes, and it manages to examine which other
mods/files are already installed, and it finishes extraction, all of
which is good. However, it doesn't pull the selected files out of their
source directories; instead it installs the entire src directory (e.g.
'11-You-Chose-This-Option') into the mod's destination folder. But hey,
it's a start!