Implement auto-extract for non-fomod mods
The ones w/ proper structure, anyway. There's supposed to be a progress
dialog for archives that take a long time to extract, but I've yet to
find one that takes long enough to cause it to trigger. Need to try a
>1GB one, I guess.

But anyway, the extraction process seems to work fine. Still need to
implement a hook to get the new mods added to the mod table.