Allow dropping items above/below other items
It was always supposed to work this way, but it didn't. Now, depending
on just where the user hovers their mouse when they release the dragged
selection, the section could be placed above or below the "parent"
item, where previously it would only ever be placed above. Also,
dragging to the empty space after the final item will move the
selection to the end of the list, as it should.