Further stabilization of the extraction process
I tore the world apart trying to find out why asyncio wasn't giving me
(closer) to real-time output from my shell subprocess...until I
realized that the problem actually seems to be that 7z block-buffers
its output (as opposed to line-buffering) when attached to a pipe, so
it just wasn't PRODUCING output for many seconds at a time...I guess
you're off the hook this time, asyncio. There's not really much I can
do about this, so I'll have to rethink how we show the progress of an
extraction.